requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
gunicorn==21.2.0
apscheduler==3.10.4
flask-mail==0.9.1
python-dotenv==1.0.0
fake-useragent==1.4.0
# Removed lxml - using html.parser instead!